requirements.txt
pandas==2.0.2
numpy==1.25.0
jupyterlab==3.6.3
matplotlib==3.7.1
seaborn==0.12.2
statsmodels==0.14.0
scikit-learn==1.2.2
pmdarima==2.0.3
streamlit
gspread
google-auth
plotly

